Remove templating on S and D
