+++ src/consensus/pbft/libbyz/node.cpp
@@ -281,7 +281,7 @@ void Node::send_to_replicas(Message* m)
   }
 }
 
-void Node::set_f(ccf::NodeId f)
+void Node::set_f(size_t f)
 {
   LOG_INFO << "***** setting f to " << f << "*****" << std::endl;
   max_faulty = f;
