+++ src/consensus/pbft/libbyz/node.h
@@ -41,7 +41,7 @@ public:
 
   size_t num_of_replicas() const;
   size_t f() const;
-  void set_f(ccf::NodeId f);
+  void set_f(size_t f);
   size_t num_correct_replicas() const;
 
   int id() const;
