+++ src/kv/msg_pack_serialise.h
@@ -5,7 +5,6 @@
 #include "../ds/msgpack_adaptor_nlohmann.h"
 #include "../ds/serialized.h"
 #include "generic_serialise_wrapper.h"
-#include "kv_types.h"
 
 #include <iterator>
 #include <msgpack/msgpack.hpp>
