+++ src/node/entities.h
@@ -68,6 +68,6 @@ namespace ccf
 
   using StoreSerialiser = kv::KvStoreSerialiser;
   using StoreDeserialiser = kv::KvStoreDeserialiser;
-  using Store = kv::Store<StoreSerialiser, StoreDeserialiser>;
-  using Tx = kv::Tx<StoreSerialiser, StoreDeserialiser>;
+  using Store = kv::Store;
+  using Tx = kv::Tx;
 }
